stop run data appending itself.
Originally I had thought to have a single data dict with all of the runs
in it. This isn't what I used in practice as it is much more flexible to
have a single dict / pickle / data file per run. To make this the actual
behaviour, I move the data = {} inside the loop that extracts data from
runs.