remove dependence on camera offset
developing a correction for a general offset between the cameras was a
bad idea. There are less than 5 runs that are out of sync. It just
happened that the canonical run used to develop the wave code was one of
those!

proc_im will be run again with all runs in sync to produce a clean data
set with no need for offset correction.

what does need modifying is the parallax_corr maths.