Output more compact.
This is as good a first release as I'm going to get. It suffers from an
excess of prints during the process.

An annoying feature is the 'fudge' factor used to make the camera data
streams mesh properly. I think this is something to do with parallax,
but I don't know what and can't seem to work it out. At least all the
other metrics to do with the cameras are defined in the top level config
file now.