fix pil problem. still problem w/ cam2 crop.
PIL wasn't cooperating because the perspective projection coefficients
need to be computed for mapping inputs to outputs, not the other way
around. This was an easy fix once spotted.

Cleaned up no longer used functions.

Changed the structure of proc_images to allow input and output dirs.

Problem with cam2 is in the transformation - chopping off the left edge
for some reason.