Parallax correction to front plane.
The maths used for parallax correction was correct, but it was
projecting to the middle plane of the tank. This causes problems because
the offsets used to normalise pixel measurements are taken from the
front plane and their projection to the middle plane is different for
each camera. Projecting to the front plane fixes this.

The issue wasn't picked up before because of a typo in the config file
where the cam2 centre position was set incorrectly. Fixing this exposed
the problem.