StitchedImage class and doc
We initialise with two input images - one from each camera. The stitched
image is a combination of the two inputs into a single image that is
aligned in physical space.

I've used generators here - they are fast. Should consider using to
replace list comps elsewhere.

The implemented method of writing out is more expressive I think. It
delegates the action of writing to disk to the Image class, which in
turn gets the path to write to from the run that it is initialised with.