synchronisation script
used to extract frames from all of the videos.

this used the whistle transient as the method of synchronisation.

i think there is still potential in using the light switch transient.

it should also be possible just to convolve the audio streams, but i
haven't made this work yet. In principle you just have to take the cross
correlation of the two signals, or the convolution with one of them
reversed (use scipy.signal.fftconvolve), whilst zero padding the signals
so that you don't cycle round.

http://dsp.stackexchange.com/questions/736/how-do-i-implement-cross-correlation-to-prove-two-audio-files-are-similar
https://gist.github.com/endolith/376572
http://stackoverflow.com/questions/7697169/padding-arrays-using-numpy