switch to allow explicit load image to memory
this is needed because for runs with lots of files we get
an error when trying to process because the number of open files exceeds
the limit set by the OS (see `ulimit -n`).

We can get around this by explicitly loading the images to memory and
then discarding the file object that they came from.

https://mail.python.org/pipermail/image-sig/2010-May/006299.html

This is sub optimal as it takes up lots of memory. It is only a problem
when using parallel_process - if this could accept a generator as input
then there would be no issue as we are using generators throughout here.