Use json file to store config used to process run
This is so that processed runs are immune to changes in config, getting
all that they need from a .json file stored with the output from initial
processing.

There was an issue with serializing numpy.int64 types, which is solved
with using .tolist() where needed.

Still need to put real <-> pixel conversion into a Processed class, so
it is dependent on the crop box from the json and not the config.